moved custommarkdownview to demo_components, removed extraneous footnote, refactored click handler in markdownsynapse
