--- src/lib/containers/CustomMarkdownErrorView.js
@@ -1,17 +1,0 @@
-
-import React from 'react'
-
-export class CustomMarkdownErrorView extends React.Component {
-
-    render () {
-        return (
-        <div className="text-danger">
-            {this.props.message}
-        </div>
-        )
-    }
-}
-
-
-
-export default CustomMarkdownErrorView
